+++ pkg/server/update-api.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+for i in types spec gorilla; do oapi-codegen --package generated --generate $i -o pkg/server/generated/$i.go pkg/server/openapi/sever.spec.yaml ; done