+++ pkg/trivy/config.go
@@ -21,10 +21,11 @@ import (
 	"github.com/eschercloudai/baski/pkg/constants"
 	"github.com/eschercloudai/baski/pkg/s3"
 	"log"
+	"strings"
 )
 
 // GenerateUserData Creates the user data that will be passed to the server being created so that a .trivyignore can be added and the scan can be run as per the users wishes.
-func GenerateUserData(s3 *s3.S3, ignoreFileName string, ignoreList []string) []byte {
+func GenerateUserData(s3 s3.InterfaceS3, ignoreFileName string, ignoreList []string) []byte {
 	trivyIgnoreData := generateTrivyFile(s3, ignoreFileName, ignoreList)
 
 	log.Println("generating userdata")
@@ -39,7 +40,8 @@ func GenerateUserData(s3 *s3.S3, ignoreFileName string, ignoreList []string) []b
 	mv ./trivy /usr/local/bin/trivy;
 fi`, constants.TrivyVersion, constants.TrivyVersion, constants.TrivyVersion, constants.TrivyVersion)
 
-	runScanData := "sudo trivy rootfs --scanners vuln -f json -o /tmp/results.json /;"
+	// Set the default command to run here - it may get overridden later.
+	runScanCommand := "sudo trivy rootfs --scanners vuln -f json -o /tmp/results.json /;"
 
 	// Prepare .trivyignore file
 	if len(trivyIgnoreData) > 0 {
@@ -49,7 +51,8 @@ cat << EOF > /tmp/.trivyignore
 EOF
 `, string(trivyIgnoreData))
 
-		runScanData = "sudo trivy rootfs --ignorefile /tmp/.trivyignore --scanners vuln -f json -o /tmp/results.json /;"
+		//Override the command to run as we now have a .trivyignore to add
+		runScanCommand = "sudo trivy rootfs --ignorefile /tmp/.trivyignore --scanners vuln -f json -o /tmp/results.json /;"
 	}
 
 	// Put it all together
@@ -59,14 +62,19 @@ touch /tmp/finished;
 %s
 %s
 echo done > /tmp/finished;
-`, trivyIgnoreFile, trivyUserData, runScanData))
+`, trivyIgnoreFile, trivyUserData, runScanCommand))
 
 }
 
 // generateTrivyFile generates the trivyignore file to be used during the scan.
-func generateTrivyFile(s3 *s3.S3, ignoreFileName string, ignoreList []string) []byte {
+func generateTrivyFile(s3 s3.InterfaceS3, ignoreFileName string, ignoreList []string) []byte {
 	var ignoreListData, trivyIgnoreFile []byte
 	var err error
+
+	//We return nothing if there are no checks required
+	if len(ignoreList) == 0 && len(ignoreFileName) == 0 {
+		return nil
+	}
 
 	// Check if a list of CVEs was passed in before checking for a trivyIgnore file
 	if len(ignoreList) != 0 {
@@ -80,16 +88,12 @@ func generateTrivyFile(s3 *s3.S3, ignoreFileName string, ignoreList []string) []
 		}
 	}
 
-	return []byte(fmt.Sprintf("%s %s", string(trivyIgnoreFile), string(ignoreListData)))
+	return []byte(fmt.Sprintf("%s\n%s", string(trivyIgnoreFile), string(ignoreListData)))
 }
 
 // parseIgnoreList turns the ignore list passed into a format that can be used in the trivyignore file.
 func parseIgnoreList(ignoreList []string) []byte {
-	var list string
-
-	for i := 0; i < len(ignoreList); i++ {
-		list = fmt.Sprintf("\n%s\n%s\n", list, ignoreList[i])
-	}
+	list := strings.Join(ignoreList, "\n")
 
 	return []byte(list)
 }